+++ src/Shambala.Repository/ReadOutgoingShipmentRepository.cs
@@ -19,7 +19,7 @@ namespace Shambala.Repository
             this.context = context;
         }
 
-        public OutgoingShipmentAggregateBLL GetDetails(long Id)
+        public OutgoingShipmentAggregateBLL GetDetails(int Id)
         {
             var outgoingShipment = context.OutgoingShipment.Include(e=>e.SalesmanIdFkNavigation).Where(e => e.Id == Id);
             var outgoingShipmentDetails = context.OutgoingShipmentDetails.Include(e => e.CustomCaratPrices)
